Compilacion 1

+++ AlbaMobileAndroid/src/main/java/es/albainformatica/albamobileandroid/comunicaciones/MiscComunicaciones.kt
@@ -2373,19 +2373,19 @@ class MiscComunicaciones(context: Context, desdeServicio: Boolean) {
                             } else {
                                 val queNumPedido = seriesDao?.getNumPedido(serieEnt.serie, queEjercicio.toInt()) ?: 0
                                 if (queNumPedido < serieEnt.pedido)
-                                    seriesDao?.setNumPedido(serieEnt.serie, queEjercicio.toInt(), queNumPedido)
+                                    seriesDao?.setNumPedido(serieEnt.serie, queEjercicio, queNumPedido)
 
                                 val queNumFra = seriesDao?.getNumFactura(serieEnt.serie, queEjercicio.toInt()) ?: 0
                                 if (queNumFra < serieEnt.factura)
-                                    seriesDao?.setNumFactura(serieEnt.serie, queEjercicio.toInt(), queNumFra)
+                                    seriesDao?.setNumFactura(serieEnt.serie, queEjercicio, queNumFra)
 
                                 val queNumAlb = seriesDao?.getNumAlbaran(serieEnt.serie, queEjercicio.toInt()) ?: 0
                                 if (queNumFra < serieEnt.albaran)
-                                    seriesDao?.setNumAlbaran(serieEnt.serie, queEjercicio.toInt(), queNumAlb)
+                                    seriesDao?.setNumAlbaran(serieEnt.serie, queEjercicio, queNumAlb)
 
                                 val queNumPres = seriesDao?.getNumPresupuesto(serieEnt.serie, queEjercicio.toInt()) ?: 0
                                 if (queNumPres < serieEnt.presupuesto)
-                                    seriesDao?.setNumPresupuesto(serieEnt.serie, queEjercicio.toInt(), queNumPres)
+                                    seriesDao?.setNumPresupuesto(serieEnt.serie, queEjercicio, queNumPres)
                             }
                         }
                     }
